+++ snakemake/sourcecache.py
@@ -13,7 +13,13 @@ import io
 from abc import ABC, abstractmethod
 
 
-from snakemake.common import is_local_file, get_appdirs, parse_uri, smart_join
+from snakemake.common import (
+    ON_WINDOWS,
+    is_local_file,
+    get_appdirs,
+    parse_uri,
+    smart_join,
+)
 from snakemake.exceptions import WorkflowError, SourceFileError
 from snakemake.io import git_content, split_git_path
 from snakemake.logging import logger
@@ -105,6 +111,9 @@ class LocalSourceFile(SourceFile):
     def simplify_path(self):
         return utils.simplify_path(self.path)
 
+    def __fspath__(self):
+        return self.path
+
 
 class LocalGitFile(SourceFile):
     def __init__(
@@ -147,12 +156,14 @@ class HostingProviderFile(SourceFile):
 
     def __init__(
         self,
-        repo: str,
-        path: str,
+        repo: str = None,
+        path: str = None,
         tag: str = None,
         branch: str = None,
         commit: str = None,
     ):
+        if repo is None:
+            raise SourceFileError("repo must be given")
         if not self.__class__.valid_repo.match(repo):
             raise SourceFileError(
                 "repo {} is not a valid repo specification (must be given as owner/name)."
@@ -197,6 +208,10 @@ class HostingProviderFile(SourceFile):
 
     def join(self, path):
         path = os.path.normpath("{}/{}".format(self.path, path))
+        if ON_WINDOWS:
+            # convert back to URL separators
+            # (win specific separators are introduced by normpath above)
+            path = path.replace("\\", "/")
         return self.__class__(
             repo=self.repo,
             path=path,
@@ -232,7 +247,7 @@ class GitlabFile(HostingProviderFile):
 
 def infer_source_file(path_or_uri, basedir: SourceFile = None):
     if isinstance(path_or_uri, SourceFile):
-        if basedir is None:
+        if basedir is None or isinstance(path_or_uri, HostingProviderFile):
             return path_or_uri
         else:
             path_or_uri = path_or_uri.get_path_or_uri()